[Ref] Splitting ONe backend into classes, pt. 5
* Added support for loading per-backend extensions
* Merged os_tpl, resource_tpl with compute
* Fixed overly complicated authentication routine
