+++ etc/backends/opennebula/model/compute/contextualization.json
