+++ lib/backends/opennebula/compute.rb
@@ -1,8 +1,9 @@
 module Backends
   module Opennebula
-    module Compute
+    class Compute < Backends::Opennebula::Base
       COMPUTE_NINTF_REGEXP = /compute_(?<compute_id>\d+)_nic_(?<compute_nic_id>\d+)/
       COMPUTE_SLINK_REGEXP = /compute_(?<compute_id>\d+)_disk_(?<compute_disk_id>\d+)/
+      OS_TPL_TERM_PREFIX = 'uuid'
 
       # Gets all compute instance IDs, no details, no duplicates. Returned
       # identifiers must correspond to those found in the occi.core.id
@@ -427,7 +428,99 @@ module Backends
         true
       end
 
+      # Returns a collection of custom mixins introduced (and specific for)
+      # the enabled backend. Only mixins and actions are allowed.
+      #
+      # @return [Occi::Collection] collection of extensions (custom mixins and/or actions)
+      def compute_get_extensions
+        read_extensions 'compute'
+      end
+
+      # Gets backend-specific `os_tpl` mixins which should be merged
+      # into Occi::Model of the server.
+      #
+      # @example
+      #    mixins = os_tpl_list #=> #<Occi::Core::Mixins>
+      #    mixins.first #=> #<Occi::Core::Mixin>
+      #
+      # @return [Occi::Core::Mixins] a collection of mixins
+      def os_tpl_list
+        os_tpl = Occi::Core::Mixins.new
+        backend_tpl_pool = ::OpenNebula::TemplatePool.new(@client)
+        rc = backend_tpl_pool.info_all
+        check_retval(rc, Backends::Errors::ResourceRetrievalError)
+
+        backend_tpl_pool.each do |backend_tpl|
+          depends = %w|http://schemas.ogf.org/occi/infrastructure#os_tpl|
+          term = os_tpl_list_tpl_to_term(backend_tpl)
+          scheme = "#{@options.backend_scheme}/occi/infrastructure/os_tpl#"
+          title = backend_tpl['NAME']
+          location = "/mixin/os_tpl/#{term}/"
+          applies = %w|http://schemas.ogf.org/occi/infrastructure#compute|
+
+          os_tpl << Occi::Core::Mixin.new(scheme, term, title, nil, depends, nil, location, applies)
+        end
+
+        os_tpl
+      end
+
+      # Gets a specific os_tpl mixin instance as Occi::Core::Mixin.
+      # Term given as an argument must match the term inside
+      # the returned Occi::Core::Mixin instance.
+      #
+      # @example
+      #    os_tpl = os_tpl_get('65d4f65adfadf-ad2f4ad-daf5ad-f5ad4fad4ffdf')
+      #        #=> #<Occi::Core::Mixin>
+      #
+      # @param term [String] OCCI term of the requested os_tpl mixin instance
+      # @return [Occi::Core::Mixin, nil] a mixin instance or `nil`
+      def os_tpl_get(term)
+        # TODO: make it more efficient!
+        os_tpl_list.to_a.select { |m| m.term == term }.first
+      end
+
+      # Gets platform- or backend-specific `resource_tpl` mixins which should be merged
+      # into Occi::Model of the server.
+      #
+      # @example
+      #    mixins = resource_tpl_list #=> #<Occi::Core::Mixins>
+      #    mixins.first  #=> #<Occi::Core::Mixin>
+      #
+      # @return [Occi::Core::Mixins] a collection of mixins
+      def resource_tpl_list
+        @resource_tpl
+      end
+
+      # Gets a specific resource_tpl mixin instance as Occi::Core::Mixin.
+      # Term given as an argument must match the term inside
+      # the returned Occi::Core::Mixin instance.
+      #
+      # @example
+      #    resource_tpl = resource_tpl_get('65d4f65adfadf-ad2f4ad-daf5ad-f5ad4fad4ffdf')
+      #        #=> #<Occi::Core::Mixin>
+      #
+      # @param term [String] OCCI term of the requested resource_tpl mixin instance
+      # @return [Occi::Core::Mixin, nil] a mixin instance or `nil`
+      def resource_tpl_get(term)
+        resource_tpl_list.to_a.select { |m| m.term == term }.first
+      end
+
       private
+
+      def os_tpl_list_tpl_to_term(tpl)
+        fixed = tpl['NAME'].downcase.gsub(/[^0-9a-z]/i, '_')
+        fixed = fixed.gsub(/_+/, '_').chomp('_').reverse.chomp('_').reverse
+        "#{OS_TPL_TERM_PREFIX}_#{fixed}_#{tpl['ID']}"
+      end
+
+      def os_tpl_list_term_to_id(term)
+        matched = term.match(/^.+_(?<id>\d+)$/)
+
+        fail Backends::Errors::IdentifierNotValidError,
+             "OsTpl term is invalid! #{term.inspect}" unless matched
+
+        matched[:id].to_i
+      end
 
       # Load methods called from compute_list/compute_get
       include Backends::Opennebula::Helpers::ComputeParseHelper
